Move Hero to its own component
