+++ src/components/store/index.ts
@@ -1,4 +1,5 @@
 /* eslint-disable import/prefer-default-export */
 export { default as Diamonds } from './Diamonds';
+export { default as Hero } from './Hero';
 export { default as ItemCard } from './ItemCard';
 export { default as Navbar } from './Navbar';
