+++ src/styles/pages/store/index.module.scss
@@ -6,179 +6,4 @@
   gap: 2rem;
   margin: 0 auto;
   max-width: 81rem;
-
-  .hero {
-    display: flex;
-    gap: 14rem;
-    min-height: 53rem;
-    overflow: hidden;
-    position: relative;
-    z-index: 0;
-
-    @media (max-width: vars.$breakpoint-lg) {
-      align-items: center;
-      flex-direction: column;
-      gap: 0;
-    }
-
-    .left {
-      margin-top: 9.5rem;
-
-      @media (max-width: vars.$breakpoint-lg) {
-        display: none;
-      }
-
-      .imageWrapper {
-        flex: none;
-        height: 26rem;
-        margin: 1.5rem;
-        position: relative;
-        width: 26rem;
-        z-index: 0;
-
-        &::before,
-        &::after {
-          border: 2px solid currentColor;
-          content: '';
-          height: 100%;
-          position: absolute;
-          width: 100%;
-          z-index: -1;
-        }
-
-        &::before {
-          left: -1.5rem;
-          top: 1.5rem;
-        }
-
-        &::after {
-          left: 1.5rem;
-          top: -1.5rem;
-        }
-
-        img {
-          object-fit: cover;
-        }
-      }
-    }
-
-    .right {
-      margin-left: auto;
-      max-width: 34.75rem;
-
-      @media (max-width: vars.$breakpoint-lg) {
-        margin: 0;
-      }
-
-      .deco1 {
-        margin-bottom: 4.75rem;
-        margin-left: -2.5rem;
-        margin-top: 2.5rem;
-        @media (max-width: vars.$breakpoint-lg) {
-          display: none;
-        }
-      }
-
-      .heroContent {
-        align-items: flex-start;
-        display: flex;
-        flex-direction: column;
-        gap: 1.5rem;
-
-        .heading {
-          color: #fff;
-          font-size: 4rem;
-          position: relative;
-          z-index: 0;
-
-          svg {
-            max-width: 100%;
-          }
-        }
-
-        .description {
-          font-size: 1.375rem;
-          line-height: 1.75rem;
-        }
-
-        .textButton {
-          background: none;
-          color: var(--theme-primary-5);
-          font: inherit;
-          padding: 0;
-          text-align: left;
-        }
-      }
-
-      .deco2 {
-        display: block;
-        margin-left: auto;
-        margin-top: 8rem;
-        @media (max-width: vars.$breakpoint-lg) {
-          display: none;
-        }
-      }
-    }
-
-    .rainbowWrapper {
-      bottom: 0;
-      display: flex;
-      left: -8rem;
-      position: absolute;
-      z-index: -1;
-
-      @media (max-width: vars.$breakpoint-lg) {
-        left: 0;
-        margin-top: -10rem;
-        position: relative;
-        transform: rotate(-25.31deg) scale(0.3);
-      }
-
-      .teal {
-        color: vars.$cyan-5;
-      }
-
-      .orange {
-        color: vars.$orange-5;
-      }
-
-      .red {
-        color: vars.$scarlet-4;
-      }
-
-      .pink {
-        color: vars.$pink-4;
-      }
-
-      .rainbow {
-        fill: none;
-        stroke: currentColor;
-        stroke-width: 24px;
-      }
-
-      .barsLeft,
-      .barsRight {
-        display: flex;
-        flex-direction: column;
-        gap: 14px;
-        position: absolute;
-        width: 100vw;
-
-        div {
-          background-color: currentColor;
-          height: 26px;
-        }
-      }
-
-      .barsLeft {
-        right: 100%;
-        top: 0;
-      }
-
-      .barsRight {
-        bottom: 0;
-        left: 100%;
-      }
-    }
-  }
 }
