+++ src/styles/pages/store/index.module.scss.d.ts
@@ -1,23 +1,5 @@
 export type Styles = {
-  barsLeft: string;
-  barsRight: string;
   container: string;
-  deco1: string;
-  deco2: string;
-  description: string;
-  heading: string;
-  hero: string;
-  heroContent: string;
-  imageWrapper: string;
-  left: string;
-  orange: string;
-  pink: string;
-  rainbow: string;
-  rainbowWrapper: string;
-  red: string;
-  right: string;
-  teal: string;
-  textButton: string;
 };
 
 export type ClassNames = keyof Styles;
